Реализация валидации данных

+++ src/main/java/com/kulachok/kulachok/service/UserServiceImpl.java
@@ -3,10 +3,11 @@ package com.kulachok.kulachok.service;
 import com.kulachok.kulachok.dto.UserDto;
 import com.kulachok.kulachok.entity.Actris;
 import com.kulachok.kulachok.entity.Cash;
-import com.kulachok.kulachok.entity.CashAccountHolder;
 import com.kulachok.kulachok.entity.Transfer;
 import com.kulachok.kulachok.entity.User;
 import com.kulachok.kulachok.entity.UserSubscription;
+import com.kulachok.kulachok.entity.model_Interface.CashAccountHolder;
+import exception.ResourceNotFoundException;
 import com.kulachok.kulachok.repository.ActrisRepository;
 import com.kulachok.kulachok.repository.CashRepository;
 import com.kulachok.kulachok.repository.TransferRepository;
@@ -14,10 +15,12 @@ import com.kulachok.kulachok.repository.UserRepository;
 import com.kulachok.kulachok.repository.UserSubscriptionRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import util.PasswordUtil;
 
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
+import java.util.NoSuchElementException;
 
 @Service
 public class UserServiceImpl implements UserService {
@@ -51,6 +54,8 @@ public class UserServiceImpl implements UserService {
         user.setUsername(userDto.getUsername());
         user.setAge(userDto.getAge());
         user.setEmail(userDto.getEmail());
+        user.setPassword(PasswordUtil.hashPassword(userDto.getPassword().toCharArray()));
+
         User savedUser = userRepository.save(user);
 
         Cash savedCash = new Cash();
@@ -63,6 +68,7 @@ public class UserServiceImpl implements UserService {
 
         Transfer savedTransfer = new Transfer();
         savedTransfer.setUser(savedUser);
+        savedTransfer.setTransferDate(LocalDateTime.now());
         savedTransfer.setDescription("При создании");
         transferRepository.save(savedTransfer);
         return savedUser;
@@ -70,11 +76,15 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public User update(int id, UserDto userDto) {
-        User user = userRepository.findById(id).get();
+        User user = userRepository.findById(id)
+                .orElseThrow(() -> new NoSuchElementException("User with id "
+                        + id + " not found"));
+
 
         user.setUsername(userDto.getUsername());
         user.setAge(userDto.getAge());
         user.setEmail(userDto.getEmail());
+        user.setPassword(userDto.getPassword());
 
         return userRepository.save(user);
     }
@@ -104,13 +114,25 @@ public class UserServiceImpl implements UserService {
     }
 
 
-    private void deleteTransfers(Actris accountHolder) {
-        List<Transfer> transfers = transferRepository.findByActris(accountHolder);
-        transferRepository.deleteAll(transfers);
-    }
+    private void deleteTransfers(CashAccountHolder accountHolder) {
+        List<Transfer> transfers;
 
-    private void deleteTransfers(User userHolder) {
-        List<Transfer> transfers = transferRepository.findByUser(userHolder);
+        if (accountHolder == null) {
+            throw new NullPointerException("Account holder is null");
+        }
+
+        if (User.class.isAssignableFrom(accountHolder.getClass())) {
+            transfers = transferRepository.findByUser((User) accountHolder);
+        } else if (Actris.class.isAssignableFrom(accountHolder.getClass())) {
+            transfers = transferRepository.findByActris((Actris) accountHolder);
+        } else {
+            throw new IllegalArgumentException("Unknown account holder type");
+        }
+
+        if (transfers.isEmpty()) {
+            throw new NoSuchElementException("Transfers not found");
+        }
+
         transferRepository.deleteAll(transfers);
     }
